Disabled Invalid FileName Tests (for Unix platforms), and readOnly test
Invalid file names can be managed by the operating system,
On Unix a file with the characters '|*?' can be created with an escape '\'.
Which makes them a valid name for that platform.

This could be revisited for writing new files however.

ReadOnly File has be manually tested on Ubuntu (via WSL).
And performs as expected.
(chmod 200 file.s, then java -jar MTT*.jar file.s, leads to an error that the file cannot be read)

And this cannot be tested for Windows platforms, so I am disabling the test altogether